Django-Server/requirements.txt
asgi-redis==1.4.3
asgiref==1.1.2
attrs==17.4.0
autobahn==17.10.1
Automat==0.6.0
channels==1.1.8
constantly==15.1.0
daphne==1.4.1
dj-database-url==0.4.2
Django==1.10
django-autoslug==1.9.3
django-extensions==1.9.8
djangorestframework==3.5.0
hyperlink==17.3.1
incremental==17.5.0
msgpack-python==0.4.8
redis==2.10.6
six==1.11.0
Twisted==17.9.0
txaio==2.8.2
typing==3.6.2
websocket-client==0.46.0
zope.interface==4.4.3
